fix: localNetGlobalSynchronizer moved to test
