+++ docs/wallet-integration-guide/examples/scripts/02-two-step-transfer/index.ts
@@ -1,8 +1,4 @@
-import {
-    localNetGlobalSynchronizer,
-    localNetStaticConfig,
-    SDK,
-} from '@canton-network/wallet-sdk'
+import { localNetStaticConfig, SDK } from '@canton-network/wallet-sdk'
 import { pino } from 'pino'
 import _accept from './_accept.js'
 import { TransferTestScriptParameters } from './types.js'
@@ -13,6 +9,7 @@ import {
     TOKEN_NAMESPACE_CONFIG,
     TOKEN_PROVIDER_CONFIG_DEFAULT,
     AMULET_NAMESPACE_CONFIG,
+    localNetGlobalSynchronizer,
 } from '../utils/index.js'
 
 const logger = pino({ name: 'v1-02-two-step-transfer', level: 'info' })
