+++ docs/wallet-integration-guide/examples/scripts/10-init-with-ledger-provider.ts
@@ -1,5 +1,4 @@
 import {
-    localNetGlobalSynchronizer,
     localNetStaticConfig,
     SDK,
     LedgerProvider,
@@ -11,6 +10,7 @@ import {
     TOKEN_NAMESPACE_CONFIG,
     TOKEN_PROVIDER_CONFIG_DEFAULT,
     AMULET_NAMESPACE_CONFIG,
+    localNetGlobalSynchronizer,
 } from './utils/index.js'
 
 import { AuthTokenProvider } from '@canton-network/core-wallet-auth'
