+++ docs/wallet-integration-guide/examples/scripts/12-subscribe-to-events.ts
@@ -2,11 +2,13 @@ import pino from 'pino'
 import {
     CompletionEvent,
     UpdateEvent,
-    localNetGlobalSynchronizer,
     localNetStaticConfig,
     SDK,
 } from '@canton-network/wallet-sdk'
-import { TOKEN_PROVIDER_CONFIG_DEFAULT } from './utils/index.js'
+import {
+    TOKEN_PROVIDER_CONFIG_DEFAULT,
+    localNetGlobalSynchronizer,
+} from './utils/index.js'
 
 const logger = pino({ name: 'v1-12-subscribe-to-events', level: 'info' })
