+++ docs/wallet-integration-guide/examples/scripts/stress/02-merge-utxos-delegate.ts
@@ -1,9 +1,5 @@
 import pino from 'pino'
-import {
-    localNetGlobalSynchronizer,
-    localNetStaticConfig,
-    SDK,
-} from '@canton-network/wallet-sdk'
+import { localNetStaticConfig, SDK } from '@canton-network/wallet-sdk'
 import path from 'path'
 import { fileURLToPath } from 'url'
 import { existsSync } from 'fs'
@@ -12,6 +8,7 @@ import {
     TOKEN_NAMESPACE_CONFIG,
     TOKEN_PROVIDER_CONFIG_DEFAULT,
     AMULET_NAMESPACE_CONFIG,
+    localNetGlobalSynchronizer,
 } from '../utils/index.js'
 import { batchTap } from './utils.js'
 import Decimal from 'decimal.js'
