+++ docs/wallet-integration-guide/examples/scripts/stress/03-acs-cache.ts
@@ -1,9 +1,8 @@
+import { localNetStaticConfig, SDK } from '@canton-network/wallet-sdk'
 import {
+    TOKEN_PROVIDER_CONFIG_DEFAULT,
     localNetGlobalSynchronizer,
-    localNetStaticConfig,
-    SDK,
-} from '@canton-network/wallet-sdk'
-import { TOKEN_PROVIDER_CONFIG_DEFAULT } from '../utils/index.js'
+} from '../utils/index.js'
 import pino from 'pino'
 
 const logger = pino({ name: 'stress-03-acs-cache', level: 'info' })
