+++ docs/wallet-integration-guide/examples/snippets/allocate-party.ts
@@ -1,7 +1,6 @@
 import {
     SDK,
     TokenProviderConfig,
-    localNetGlobalSynchronizer,
     localNetStaticConfig,
 } from '@canton-network/wallet-sdk'
 
@@ -34,7 +33,7 @@ export default async function () {
     const sdk = await SDK.create({
         auth,
         ledgerClientUrl: localNetStaticConfig.LOCALNET_APP_USER_LEDGER_URL,
-        synchronizerId: localNetGlobalSynchronizer,
+        synchronizerId: global.LOCALNET_GLOBAL_SYNCHRONIZER,
     })
 
     const key = sdk.keys.generate()
