+++ docs/wallet-integration-guide/examples/snippets/compute-transaction-hash.ts
@@ -1,15 +1,11 @@
-import {
-    SDK,
-    localNetGlobalSynchronizer,
-    localNetStaticConfig,
-} from '@canton-network/wallet-sdk'
+import { SDK, localNetStaticConfig } from '@canton-network/wallet-sdk'
 
 export default async function () {
     // it is important to configure the SDK correctly else you might run into connectivity or authentication issues
     const sdk = await SDK.create({
         auth: global.TOKEN_PROVIDER_CONFIG_DEFAULT,
         ledgerClientUrl: localNetStaticConfig.LOCALNET_APP_USER_LEDGER_URL,
-        synchronizerId: localNetGlobalSynchronizer,
+        synchronizerId: global.LOCALNET_GLOBAL_SYNCHRONIZER,
     })
 
     const transaction = global.PREPARED_TRANSACTION
