+++ docs/wallet-integration-guide/examples/snippets/execute-transaction.ts
@@ -1,6 +1,5 @@
 import {
     SDK,
-    localNetGlobalSynchronizer,
     localNetStaticConfig,
     signTransactionHash,
 } from '@canton-network/wallet-sdk'
@@ -10,7 +9,7 @@ export default async function () {
     const sdk = await SDK.create({
         auth: global.TOKEN_PROVIDER_CONFIG_DEFAULT,
         ledgerClientUrl: localNetStaticConfig.LOCALNET_APP_USER_LEDGER_URL,
-        synchronizerId: localNetGlobalSynchronizer,
+        synchronizerId: global.LOCALNET_GLOBAL_SYNCHRONIZER,
     })
     const myParty = global.EXISTING_PARTY_1
     const keys = global.EXISTING_PARTY_1_KEYS
