+++ docs/wallet-integration-guide/examples/snippets/token-extended.ts
@@ -1,15 +1,11 @@
-import {
-    SDK,
-    localNetGlobalSynchronizer,
-    localNetStaticConfig,
-} from '@canton-network/wallet-sdk'
+import { SDK, localNetStaticConfig } from '@canton-network/wallet-sdk'
 
 export default async function () {
     // Create basic SDK first
     const basicSDK = await SDK.create({
         auth: global.TOKEN_PROVIDER_CONFIG_DEFAULT,
         ledgerClientUrl: localNetStaticConfig.LOCALNET_APP_USER_LEDGER_URL,
-        synchronizerId: localNetGlobalSynchronizer,
+        synchronizerId: global.LOCALNET_GLOBAL_SYNCHRONIZER,
     })
 
     // Extend with token namespace when needed
