+++ docs/wallet-integration-guide/examples/snippets/transaction-update-by-id.ts
@@ -1,14 +1,10 @@
-import {
-    SDK,
-    localNetGlobalSynchronizer,
-    localNetStaticConfig,
-} from '@canton-network/wallet-sdk'
+import { SDK, localNetStaticConfig } from '@canton-network/wallet-sdk'
 
 export default async function () {
     const sdk = await SDK.create({
         auth: global.TOKEN_PROVIDER_CONFIG_DEFAULT,
         ledgerClientUrl: localNetStaticConfig.LOCALNET_APP_USER_LEDGER_URL,
-        synchronizerId: localNetGlobalSynchronizer,
+        synchronizerId: global.LOCALNET_GLOBAL_SYNCHRONIZER,
         amulet: global.AMULET_NAMESPACE_CONFIG,
         token: global.TOKEN_NAMESPACE_CONFIG,
     })
