+++ docs/wallet-integration-guide/examples/snippets/validator-party.ts
@@ -1,7 +1,6 @@
 import {
     SDK,
     getValidatorParty,
-    localNetGlobalSynchronizer,
     localNetStaticConfig,
 } from '@canton-network/wallet-sdk'
 
@@ -9,7 +8,7 @@ export default async function () {
     const sdk = await SDK.create({
         auth: TOKEN_PROVIDER_CONFIG_DEFAULT,
         ledgerClientUrl: localNetStaticConfig.LOCALNET_APP_USER_LEDGER_URL,
-        synchronizerId: localNetGlobalSynchronizer,
+        synchronizerId: global.LOCALNET_GLOBAL_SYNCHRONIZER,
         token: {
             registries: [
                 new URL('http://localhost:2000/api/validator/v0/scan-proxy'),
