+++ sdk/wallet-sdk/src/config.ts
@@ -1,8 +1,6 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import type { SynchronizerSelector } from './wallet/init/synchronizer.js'
-
 const LOCALNET_APP_VALIDATOR_URL = new URL(
     'http://localhost:2000/api/validator'
 )
@@ -31,27 +29,4 @@ export const localNetStaticConfig = {
     LOCALNET_SV_LEDGER_URL,
     LOCALNET_TOKEN_STANDARD_URL,
     LOCALNET_USER_ID,
-}
-
-/**
- * `synchronizerId` selector for `SDK.create` that picks LocalNet's global
- * synchronizer. LocalNet can additionally run an app-synchronizer, and the SDK
- * refuses to guess between them — pass this when your code targets the global one.
- */
-export const localNetGlobalSynchronizer: SynchronizerSelector = (
-    synchronizers
-) => {
-    const global = synchronizers.find(
-        (s) =>
-            s.synchronizerAlias === 'global' ||
-            s.synchronizerAlias === 'global-domain'
-    )
-    if (!global) {
-        throw new Error(
-            `No global synchronizer among: ${synchronizers
-                .map((s) => s.synchronizerAlias)
-                .join(', ')}`
-        )
-    }
-    return global.synchronizerId
 }
